ipmiparm.cpp: Bug fixes
Well, I thought there were no bugs in V1.1, but I hadn't tested it on
real hardware, yet. There is one ipmi parameter, "hotmod", that gives a
permission denied error when the program tries to access it. After that
the program cannot read anything out of that directory. There's a hack
in the code for now to skip the hotmod parameter, but I'll research it
so I can correctly handle permission errors in the future.