ipmiparm.cpp - int and str fixes
Wasn't able to detect whether the contents of a kmod parameter were
string or integer. Added some checking to make that determination.
Also tightened the requirements for an int, so foo would not be
accepted as 0xf (15).

You can change a parameter from a string to an int, but you cannot
change from an int to a string. This is to protect the user from
shooting himself in the foot with invalid entries.  Changing from a
string to an int is allowed, because it is possible for a parameter that
is supposed to be an int to have a null string.